FIX: Maintain backwards compatibility
