+++ sdk/csharp/libraries/microsoft.bot.builder.solutions/Skills/SkillControllerBase.cs
@@ -1,0 +1,135 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.IO;
+using System.Net;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Bot.Builder.Integration.AspNet.Core;
+using Microsoft.Bot.Builder.Solutions.Skills.Auth;
+using Newtonsoft.Json;
+
+namespace Microsoft.Bot.Builder.Solutions.Skills
+{
+    /// <summary>
+    /// This is the default Controller that contains APIs for handling
+    /// calls from a channel and calls from a parent bot (to a skill bot).
+    /// </summary>
+    public abstract class SkillControllerBase : ControllerBase
+    {
+        private readonly IBot _bot;
+        private readonly IBotFrameworkHttpAdapter _botFrameworkHttpAdapter;
+        private readonly ISkillWebSocketAdapter _skillWebSocketAdapter;
+        private readonly IAuthenticator _authenticator;
+        private readonly BotSettingsBase _botSettings;
+        private readonly JsonSerializer _jsonSerializer = JsonSerializer.Create(Serialization.Settings);
+
+        public SkillControllerBase(
+            IBot bot,
+            BotSettingsBase botSettings,
+            IBotFrameworkHttpAdapter botFrameworkHttpAdapter,
+            ISkillWebSocketAdapter skillWebSocketAdapter,
+            IAuthenticator authenticator)
+        {
+            _bot = bot ?? throw new ArgumentNullException(nameof(IBot));
+            _botSettings = botSettings ?? throw new ArgumentNullException(nameof(botSettings));
+            _botFrameworkHttpAdapter = botFrameworkHttpAdapter ?? throw new ArgumentNullException(nameof(IBotFrameworkHttpAdapter));
+            _skillWebSocketAdapter = skillWebSocketAdapter;
+            _authenticator = authenticator;
+        }
+
+        // Each skill provides a template manifest file which we use to fill in the dynamic elements.
+        // There are protected to enable unit tests to mock.
+        protected HttpClient HttpClient { get; set; } = new HttpClient();
+
+        protected string ManifestTemplateFilename { get; set; } = @".\Skills\manifestTemplate.json";
+
+        /// <summary>
+        /// This API is the endpoint for when a bot receives a message from a channel or a parent bot.
+        /// </summary>
+        /// <returns>Task.</returns>
+        [Route("api/messages")]
+        [HttpPost]
+        public async Task BotMessageAsync()
+        {
+            await _botFrameworkHttpAdapter.ProcessAsync(Request, Response, _bot).ConfigureAwait(false);
+        }
+
+        /// <summary>
+        /// This API is the endpoint the bot exposes as skill.
+        /// </summary>
+        /// <returns>Task.</returns>
+        [Route("api/skill/messages")]
+        [HttpGet]
+        public async Task SkillMessageWebSocketAsync()
+        {
+            if (_skillWebSocketAdapter != null)
+            {
+                await _skillWebSocketAdapter.ProcessAsync(Request, Response, _bot).ConfigureAwait(false);
+            }
+            else
+            {
+                Response.StatusCode = (int)HttpStatusCode.MethodNotAllowed;
+            }
+        }
+
+        /// <summary>
+        /// This API is the manifest endpoint that surfaces a self-describing document of the Skill.
+        /// The base template is provided by the template and at runtime we fill in endpoint, msaAppId
+        /// and as requested the triggering utterances inline within the manifest.
+        /// </summary>
+        /// <param name="inlineTriggerUtterances">Include triggering utterances inline in manifest.</param>
+        /// <returns>Task.</returns>
+        [Route("api/skill/manifest")]
+        [HttpGet]
+        public async Task SkillManifestAsync([Bind, FromQuery] bool inlineTriggerUtterances = false)
+        {
+            try
+            {
+                string skillUriBase = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host}";
+
+                SkillManifestGenerator manifestGenerator = new SkillManifestGenerator(HttpClient);
+                var skillManifest = await manifestGenerator.GenerateManifestAsync(ManifestTemplateFilename, _botSettings.MicrosoftAppId, _botSettings.CognitiveModels, skillUriBase, inlineTriggerUtterances).ConfigureAwait(false);
+
+                Response.ContentType = "application/json";
+                Response.StatusCode = 200;
+
+                using (var writer = new StreamWriter(Response.Body))
+                {
+                    using (var jsonWriter = new JsonTextWriter(writer))
+                    {
+                        _jsonSerializer.Serialize(jsonWriter, skillManifest);
+                    }
+                }
+            }
+            catch (Exception e)
+            {
+                Response.ContentType = "application/json";
+                Response.StatusCode = (int)HttpStatusCode.InternalServerError;
+
+                using (var writer = new StreamWriter(Response.Body))
+                {
+                    using (var jsonWriter = new JsonTextWriter(writer))
+                    {
+                        _jsonSerializer.Serialize(jsonWriter, e.Message);
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// This API is for calling bots to see if a skill is live
+        /// Additional it can also be used to verify if calling bot can pass authentication
+        /// because this API verifies the bearer token in the process.
+        /// </summary>
+        /// <returns>Task.</returns>
+        [Route("api/skill/ping")]
+        [HttpGet]
+        public async Task SkillPingAsync()
+        {
+            await _authenticator.AuthenticateAsync(Request, Response).ConfigureAwait(false);
+        }
+    }
+}
